# app/utils/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "FitHub - API",
        "description": "Welcome to FitHub api's documentation. Here you will find all the routes at disposal including details and their HTTP methods."
    },
...
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Fetch all users & their public informations.",
                "description": "Returns an array of objects with users' informations.",
                "operationid": "findAll",
                "responses": {
                    "200": {
                        "description": "OK - Returns a JSON object users' data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User cannot be found."
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user account.",
                "description": "Create a new user account based on the data retrieved.",
                "operationId": "createOne",
                "requestBody": {
                    "description": "Add a new user in FitHub.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                }
            }
        },
...
        "/activity/user": {
            "post": {
                "tags": [
                    "Activity"
                ],
                "summary": "Assign an activity to a user.",
                "description": "Assign an activity",
                "operationId": "assignActivityToUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "User's id to assign an activity.",
                        "required": "true"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "description": "Activities' data.",
                            "schema" : {
                                "$ref": "#/components/schemas/Activity"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK - Activity assigned to user !"
                    },
                    "404": {
                        "description": "Activity or user not found."
                    }
                }
            }
        },
        "/activity/:userId/:activityUserId": {
            "delete": {
                "tags": [
                    "Activity"
                ],
                "summary": "Removed an activity from a user.",
                "description": "Removed an activity from a user.",
                "operationId": "removeActivityFromUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User's id to remove an activity.",
                        "required": "true"
                    },
                    {
                        "name": "activityUserId",
                        "in": "path",
                        "description": "User's activity and date entry.",
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Activity removed from user !"
                    },
                    "404": {
                        "description": "Activity or user not found."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "firstname": {
                        "type": "string",
                        "example": "John",
                        "required": "true"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Doe",
                        "required": "true"
                    },
                    "nickname": {
                        "type": "string",
                        "example": "John_Doe_73",
                        "required": "true"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0123456789",
                        "required": "false",
                        "default": "null"
                    },
                    "role": {
                        "type": "string",
                        "example": "user",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "default": "user"
                    },
                    "login_streak": {
                        "type": "integer",
                        "example": 10,
                        "default": 0
                    },
                    "password": {
                        "type": "string",
                        "example": "$2b$10$wyzLD2T8Jxpb2Fp4r7cgWOapKwfTGkLGMKq1VVwsc2r4LTNUDXIhq",
                        "required": "true"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe73@mymail.com",
                        "required": "true"
                    },
                    "weight": {
                        "type": "integer",
                        "example": 64,
                        "required": "true"
                    },
                    "age": {
                        "type": "integer",
                        "example": 25,
                        "required": "true"
                    },
                    "gender": {
                        "type": "string",
                        "example": "homme",
                        "enum": [
                            "femme",
                            "homme",
                            "non-spécifié"
                        ],
                        "default": "non-spécifié"
                    },
                    "challenge_id": {
                        "type": "integer",
                        "example": 2,
                        "default": "null"
                    },
                    "xp": {
                        "type": "integer",
                        "example": 478054,
                        "default": "null"
                    },
                    "profile_visibility": {
                        "type": "string",
                        "example": "privé",
                        "enum": [
                            "publique",
                            "privé"
                        ],
                        "default": "publique"
                    },
                    "image_path": {
                        "type": "string",
                        "example": "dirName/fileName123",
                        "default": "null"
                    },
                    "image_mimetype": {
                        "type": "string",
                        "example": "image/jpeg",
                        "default": "null"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "2023-04-02 13:39:13.472+02"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "null"
                    }
                }
            },
...
            "Activity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "code": {
                        "type": "string",
                        "required": "true"
                    },
                    "met": {
                        "type": "number",
                        "format": "float",
                        "required": "true"
                    },
                    "label": {
                        "type": "string",
                        "required": "true"
                    },
                    "category_activity_id": {
                        "type": "integer",
                        "required": "true"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "2023-04-02 13:39:13.472+02"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "null"
                    }
                }
            }
        }
    }
}